Extract function prototypes into nss.h
