+++ src/nss.c
@@ -33,54 +33,7 @@
 
 #include "avahi.h"
 #include "util.h"
-
-#if defined(NSS_IPV4_ONLY) && !defined(MDNS_MINIMAL)
-#define _nss_mdns_gethostbyname4_r _nss_mdns4_gethostbyname4_r
-#define _nss_mdns_gethostbyname3_r _nss_mdns4_gethostbyname3_r
-#define _nss_mdns_gethostbyname2_r _nss_mdns4_gethostbyname2_r
-#define _nss_mdns_gethostbyname_r _nss_mdns4_gethostbyname_r
-#define _nss_mdns_gethostbyaddr_r _nss_mdns4_gethostbyaddr_r
-#elif defined(NSS_IPV4_ONLY) && defined(MDNS_MINIMAL)
-#define _nss_mdns_gethostbyname4_r _nss_mdns4_minimal_gethostbyname4_r
-#define _nss_mdns_gethostbyname3_r _nss_mdns4_minimal_gethostbyname3_r
-#define _nss_mdns_gethostbyname2_r _nss_mdns4_minimal_gethostbyname2_r
-#define _nss_mdns_gethostbyname_r _nss_mdns4_minimal_gethostbyname_r
-#define _nss_mdns_gethostbyaddr_r _nss_mdns4_minimal_gethostbyaddr_r
-#elif defined(NSS_IPV6_ONLY) && !defined(MDNS_MINIMAL)
-#define _nss_mdns_gethostbyname4_r _nss_mdns6_gethostbyname4_r
-#define _nss_mdns_gethostbyname3_r _nss_mdns6_gethostbyname3_r
-#define _nss_mdns_gethostbyname2_r _nss_mdns6_gethostbyname2_r
-#define _nss_mdns_gethostbyname_r _nss_mdns6_gethostbyname_r
-#define _nss_mdns_gethostbyaddr_r _nss_mdns6_gethostbyaddr_r
-#elif defined(NSS_IPV6_ONLY) && defined(MDNS_MINIMAL)
-#define _nss_mdns_gethostbyname4_r _nss_mdns6_minimal_gethostbyname4_r
-#define _nss_mdns_gethostbyname3_r _nss_mdns6_minimal_gethostbyname3_r
-#define _nss_mdns_gethostbyname2_r _nss_mdns6_minimal_gethostbyname2_r
-#define _nss_mdns_gethostbyname_r _nss_mdns6_minimal_gethostbyname_r
-#define _nss_mdns_gethostbyaddr_r _nss_mdns6_minimal_gethostbyaddr_r
-#elif defined(MDNS_MINIMAL)
-#define _nss_mdns_gethostbyname4_r _nss_mdns_minimal_gethostbyname4_r
-#define _nss_mdns_gethostbyname3_r _nss_mdns_minimal_gethostbyname3_r
-#define _nss_mdns_gethostbyname2_r _nss_mdns_minimal_gethostbyname2_r
-#define _nss_mdns_gethostbyname_r _nss_mdns_minimal_gethostbyname_r
-#define _nss_mdns_gethostbyaddr_r _nss_mdns_minimal_gethostbyaddr_r
-#endif
-
-// Define prototypes for nss function we're going to export (fixes GCC warnings)
-#ifndef __FreeBSD__
-enum nss_status _nss_mdns_gethostbyname4_r(const char*, struct gaih_addrtuple**,
-                                           char*, size_t, int*, int*, int32_t*);
-#endif
-enum nss_status _nss_mdns_gethostbyname3_r(const char*, int, struct hostent*,
-                                           char*, size_t, int*, int*, int32_t*,
-                                           char**);
-enum nss_status _nss_mdns_gethostbyname2_r(const char*, int, struct hostent*,
-                                           char*, size_t, int*, int*);
-enum nss_status _nss_mdns_gethostbyname_r(const char*, struct hostent*, char*,
-                                          size_t, int*, int*);
-enum nss_status _nss_mdns_gethostbyaddr_r(const void*, int, int,
-                                          struct hostent*, char*, size_t, int*,
-                                          int*);
+#include "nss.h"
 
 static avahi_resolve_result_t do_avahi_resolve_name(int af, const char* name,
                                                     userdata_t* userdata) {
